requirements.txt
# This file just points to setup.cfg, add requirements there
-e .

xmltodict==0.13.0
python-dateutil==2.8.2
requirements-parser==0.5.0